Add query string for individual types
For each endpoint, optional query strings are now accepted that conform to the key of the type. For example, for items, a query string of ?type=origin will return something like:

{"origin":["Elven"],"condition":["None"],"weapon":["None"],"armor":["None"],"healing":["None"],"mundane":["None"],"spellEffect":["None"]}

Ultimately I can probably simplify this in the get{Type}() functions so it doesn't return the blanks, but this gives the necessary functionality for now.